Match functions in item, inventory, partner, model, entity_text, evl

+++ include/dw/clock.h
@@ -20,7 +20,7 @@ void addClock(void);
 void tickGameClock(int32_t instanceId);
 void renderGameClock(int32_t instanceId);
 void updatePlaytime(int32_t instanceId);
-void advanceToTime(int32_t hour, int32_t minute);
+void advanceToTime(int32_t hour, int16_t minute);
 void updateMinuteHand(int32_t hour, int32_t minute);
 void stopGameTime(void);
 void startGameTime(void);
